Add files via upload
A Jupyter Notebook version of the code. It contains tester functions and functions that sequentially run parts of flypy and plot the errors. It imports the fly class from flypy; making classes in IPython really isn't worth it. New functionality to write the output as a .csv file will be added soon.