Add files via upload
This is a Jupyter Notebook code to sequentially test all the files in a fly for errors and plot them. Improved functionality (to write the processed error data to a .csv file) will soon be added. 
This program imports the fly() class from flypy.  It made no sense to rewrite flypy in a Jupyter format  because flypy primarily consists of class definitions.